pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "motor>=3.7.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
